pyproject.toml
[project]
name = "pycon24llms"
version = "0.1.0"
description = "Material del taller GenAI loves f-string impartido en PyConES 2024"
authors = [
    { name="Alejandro Vidal", email="ql20depbo@mozmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "bs4>=0.0.2",
    "chroma>=0.2.0",
    "langchain-chroma>=0.1.4",
    "langchain-community>=0.3.1",
    "langchain-experimental>=0.3.2",
    "langchain-openai>=0.2.1",
    "langchain>=0.3.2",
    "termcolor>=2.5.0",
]
